Remove toc.not_readable suppression (name collision), add src/pixelator/common/report/__init__.py
