+++ src/pixelator/common/report/__init__.py
@@ -1,0 +1,4 @@
+"""Common report utilities for Pixelator.
+
+Copyright © 2026 Pixelgen Technologies AB.
+"""
